% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGFRsmooth.R
\name{eGFRsmooth}
\alias{eGFRsmooth}
\title{Smooths eGFR trajectories and outputs plots for each subject, and a table
summarizing slope and percentage change for each trajectory}
\usage{
eGFRsmooth(
  data,
  k.width = 150,
  w.width = 2,
  min.width = 1,
  thres.acute = 30,
  max.range = NA,
  sort.param = "Percentage.Drop",
  decreasing = FALSE,
  output.dir = NULL
)
}
\arguments{
\item{data}{A data.frame. Input data with subject ID, numeric date and eGFR}

\item{k.width}{A numeric value. Smoothing bandwidth parameter}

\item{w.width}{A numeric value. Width of window (in years). Defaults to 
2 years}

\item{min.width}{A numeric value. Minimal width (in years) for a trajectory 
to be smoothed. Defaults to 1 year}

\item{thres.acute}{A numeric value. This argument defines the minimum width 
a data point must be away from both of it's neighbours to be kept. 
Otherwise, the data point is trimmed. This removes dense clusters of points
usually characterized by large variance. Defaults to 30 days}

\item{max.range}{A numeric value. The maximum range (in years) from the
baseline date to be analyzed. Defaults to NA (unrestricted range)}

\item{sort.param}{A character vector. Controls output plot name. One of 
Slope, Percentage.Drop, eGFR.Window.Start, eGFR.Window.End, 
eGFR.Baseline, Observations.Total, Observations.Used, Time.Range, 
Time.Window.Start, Time.Window.End.
Defaults to Percentage.Drop}

\item{decreasing}{A boolean value. Controls sorting}

\item{output.dir}{A character value. Output directory for plots and 
summary table. Defaults to NULL}
}
\value{
A data.frame. Summarizes the slope, percentage change and other 
 statistics for each trajectory. Outputs a plot (pdf) and table (txt) to 
 user defined working directory or the present working directory
}
\description{
Smooths eGFR trajectories and outputs plots for each subject, and a table
summarizing slope and percentage change for each trajectory
}
